Remove legacy IVSolver and migrate to IVSolverFDM
Remove deprecated iv_solver.hpp/cpp which duplicated IVSolverFDM implementation.
Create new BatchIVSolver wrapper that uses IVSolverFDM::solve_batch_impl() for
parallel batch processing.

Changes:
- Removed src/option/iv_solver.{hpp,cpp} (~220 lines of duplicate code)
- Created src/option/batch_iv_solver.hpp (header-only BatchIVSolver)
- Updated benchmarks to use IVSolverFDM instead of IVSolver
- Removed iv_solver target from BUILD files

Benefits:
- Eliminates 80% code duplication between IVSolver and IVSolverFDM
- Unified API: All FDM-based IV solving uses IVSolverFDM
- BatchIVSolver now uses modern IVSolverBase interface
- Stateless design (query passed to solve(), not stored in constructor)
- Leverages IVSolverFDM::solve_batch_impl() for OpenMP parallelization

Architecture after cleanup:
- IVSolverBase: Abstract base with deducing this
- IVSolverFDM: FDM-based IV solver (~143ms/query, ground truth)
- IVSolverInterpolated: B-spline IV solver (~7.5µs/query, production)
- BatchIVSolver: Convenience wrapper for batch operations

All tests pass.
