--- src/option/iv_solver.hpp
@@ -1,190 +1,0 @@
-#pragma once
-
-#include "src/math/root_finding.hpp"
-#include "common/ivcalc_trace.h"
-#include <expected>
-#include "src/support/error_types.hpp"
-#include "src/support/parallel.hpp"
-#include "src/option/iv_types.hpp"
-#include "src/option/option_spec.hpp"  // For IVQuery
-#include <optional>
-#include <string>
-#include <vector>
-#include <span>
-
-namespace mango {
-
-/// Configuration for implied volatility solver
-///
-/// Controls both the root-finding algorithm and the PDE grid
-/// used for American option pricing during IV calculation.
-struct IVConfig {
-    /// Root-finding configuration (Brent's method parameters)
-    RootFindingConfig root_config;
-
-    /// Number of spatial grid points for PDE solver
-    size_t grid_n_space = 101;
-
-    /// Number of time steps for PDE solver
-    size_t grid_n_time = 1000;
-
-    /// Maximum spot price for grid (S_max)
-    double grid_s_max = 200.0;
-};
-
-/// Implied Volatility Solver for American Options
-///
-/// Finds the volatility parameter that makes the American option's
-/// theoretical price (from PDE solver) match the observed market price.
-///
-/// **Algorithm:**
-/// Uses Brent's method for root-finding, with each iteration solving
-/// the American option PDE for a candidate volatility. This nested
-/// approach is robust but relatively slow (~250ms per IV calculation).
-///
-/// **Usage:**
-/// ```cpp
-/// IVQuery query{
-///     .option = OptionSpec{
-///         .spot = 100.0,
-///         .strike = 100.0,
-///         .maturity = 1.0,
-///         .rate = 0.05,
-///         .dividend_yield = 0.0,
-///         .type = OptionType::PUT
-///     },
-///     .market_price = 10.45
-/// };
-///
-/// IVConfig config{
-///     .root_config = RootFindingConfig{.max_iter = 100, .tolerance = 1e-6}
-/// };
-///
-/// IVSolver solver(query, config);
-/// IVResult result = solver.solve();
-///
-/// if (result.converged) {
-///     std::cout << "IV: " << result.implied_vol << "\n";
-/// }
-/// ```
-///
-/// **Performance:**
-/// - FDM-based: ~250ms per calculation (ground truth)
-/// - For production use cases requiring many queries, consider
-///   precomputed price table with interpolation (~7.5µs per query)
-///
-/// **USDT Tracing:**
-/// Emits MODULE_IMPLIED_VOL traces for monitoring:
-/// - algo_start: IV calculation begins
-/// - algo_progress: Root-finding progress
-/// - algo_complete: IV calculation completes
-/// - validation_error: Input validation failures
-/// - convergence_failed: Non-convergence diagnostics
-class IVSolver {
-public:
-    /// Construct solver with problem parameters and configuration
-    ///
-    /// @param query IV query (option specification and market price)
-    /// @param config Solver configuration (root-finding and grid settings)
-    explicit IVSolver(const IVQuery& query, const IVConfig& config);
-
-    /// Solve for implied volatility
-    ///
-    /// Uses Brent's method to find the volatility that makes the
-    /// American option's theoretical price match the market price.
-    ///
-    /// **Note:** This is a STUB implementation that returns "Not implemented"
-    /// error. Full implementation will be added in subsequent tasks.
-    ///
-    /// @return IVResult with convergence status and implied volatility
-    IVResult solve();
-
-private:
-    IVQuery query_;
-    IVConfig config_;
-    mutable std::optional<SolverError> last_solver_error_;
-
-    /// Estimate upper bound for volatility search using intrinsic value approximation
-    /// @return Upper bound estimate (typically 2.0-3.0 for reasonable markets)
-    double estimate_upper_bound() const;
-
-    /// Estimate lower bound for volatility search
-    /// @return Lower bound (typically 0.01 or 1%)
-    double estimate_lower_bound() const;
-
-    /// Objective function for root-finding: f(σ) = V(σ) - V_market
-    /// @param volatility Candidate volatility
-    /// @return Difference between theoretical and market price
-    double objective_function(double volatility) const;
-};
-
-/// Batch Implied Volatility Solver
-///
-/// Solves implied volatility for multiple options in parallel using OpenMP.
-/// This is significantly faster than solving options sequentially.
-///
-/// Example usage:
-/// ```cpp
-/// std::vector<IVQuery> batch = { ... };
-/// IVConfig config;  // Shared configuration
-///
-/// auto results = solve_implied_vol_batch(batch, config);
-/// ```
-///
-/// Performance:
-/// - Single-threaded: ~7 IVs/sec (101x1000 grid)
-/// - Parallel (32 cores): ~107 IVs/sec (15.3x speedup)
-///
-/// Use cases:
-/// - Volatility surface construction: Calculate IV for entire grid of strikes/maturities
-/// - Market data processing: Batch-process option chains
-/// - Risk calculations: Compute sensitivities across multiple scenarios
-/// - Model calibration: Evaluate objective function for optimization
-class BatchIVSolver {
-public:
-    /// Solve implied volatility for a batch of options in parallel
-    ///
-    /// @param queries Vector of IV queries (option specs and market prices)
-    /// @param config Shared configuration (grid size, tolerances)
-    /// @return Vector of IV results (same order as input)
-    static std::vector<IVResult> solve_batch(
-        std::span<const IVQuery> queries,
-        const IVConfig& config)
-    {
-        std::vector<IVResult> results(queries.size());
-
-        MANGO_PRAGMA_PARALLEL_FOR
-        for (size_t i = 0; i < queries.size(); ++i) {
-            IVSolver solver(queries[i], config);
-            results[i] = solver.solve();
-        }
-
-        return results;
-    }
-
-    /// Solve implied volatility for a batch of options (vector overload)
-    static std::vector<IVResult> solve_batch(
-        const std::vector<IVQuery>& queries,
-        const IVConfig& config)
-    {
-        return solve_batch(std::span{queries}, config);
-    }
-};
-
-/// Convenience function for batch IV solving
-inline std::vector<IVResult> solve_implied_vol_batch(
-    std::span<const IVQuery> queries,
-    const IVConfig& config)
-{
-    return BatchIVSolver::solve_batch(queries, config);
-}
-
-/// Convenience function for batch IV solving (vector overload)
-inline std::vector<IVResult> solve_implied_vol_batch(
-    const std::vector<IVQuery>& queries,
-    const IVConfig& config)
-{
-    return BatchIVSolver::solve_batch(queries, config);
-}
-
-}  // namespace mango
